In this commit...
comments.pl
-----------

-  All instances of '-1' and '5' when it comes to comment
scores have been changed to $I{comment_minscore} and
$I{comment_maxscore} respectively. The appropriate entries have been
entered into slashdotrc.pl

- Multiple moderation on one comment by the same user is now disallowed.


admin.pl
--------

- When editing a story, that story's title now appears in the brower's
titlebar.


index.pl
--------

- Section articles appearing on main page now have an ADDITIONAL link to
the section's homepage on the "Read More" line.


submit.pl
---------

- When editing a submission, that submission's title (and eventually,
the name of the story it will become) now appears in the browser's
titlebar. This is useful for admins who open lots of windows at once
when dealing with submissions and like to cover most of the window or
roll it up and yet still see what window belongs to what operation.


Slash.pm
--------

- Moderation used to be allowed if an author was logged in as an AUTHOR but
not as a user. Now moderation form elements will NOT show for any UID
that is not positive.