Minor changes to the Slashcode dearchive script as a result of regression testing. The regular expressions in the utility will work on comment .shtml files all the way back to Slash 0.9.2, however the field extraction may still need more fine tuning.
MAIN2bender: Converts a dump from a MAIN (as of 1.0.11) based Slash site
and will output a dump of the data made compatible for Bender.

The file included in the sql/updates is a metafile that contains any
specific SQL commands to update the tables as well as data which the
script uses to determine tables still in use and tables that are
obsolete (but still may need special processing involving other
existing tables). Obsoleted tables are removed from the output dump.