Fixes for several logic bugs (and the misuse of DBI::errstr as a scalar, it's not...it's a method...I keep forgetting).
The code still is not working. Something critical is affecting the
importDump() routine and is not causing it to process the lines. It's
gotta be a regexp error or something like that, but I haven't had time
to figure it out. Once this bug is killed, then this script should be
close to working...although I'm not going to rule out another showstopper
or two.