This patch fixes three things.
First, the useless timezone offset in the admin.pl title is changed
to provide more info.  (Though if the rest of the title is too long,
that info is scaled back a little.)  This adds one "SELECT NOW()" to
each admin.pl access, but it's worth it for authors to be able to see
what time the DB thinks it is.

Second is the thing I'm not quite sure about.  There was an apparent bug
with @$extracolumns -- I was seeing "Can't use an undefined value as an
ARRAY reference" so I changed the emptiness condition for $extracolumns
from an annoying undef, to a reference to an empty array.  I don't
believe this will affect any of the logic.

Third, hyphenizing the "dept" is made a little smarter so even Katz
can't screw it up.  :)