Better handling of installation when no replacement is necessary. Basically we install all perl files (with optional replacement of the invocation line(s)) using the loop and then install everything else afterwards.
It does work better this way, especially since I've deleted the lines that
normally install scripts in bin/ and sbin/ and it was silly to reimplement
them when I could just make the logic more intelligent.

Of course, it would have been simpler to re-implement the lines already
deleted, but this route gives package maintaners more flexibility.