Added - Static Makefile in the plugins/ directory that implements one target: clean. This is used by the top level Makefile.
Actually, I would recomment that all plugin compilation and installation
functionality should be moved into such to keep the top-level Makefile
clean. We're getting enough plugins where this change is warranted.

Makefile (and Makefile.NEW) now use SLASH_PREFIX to determine where
Slash will be installed on a system. This is to prevent conflicts
with any Perl modules that are installed.

Makefile.NEW now implements a (hopefully simpler) installation method for
plugins. I would appreciate it if folks would test it out and get back
to me on it works out.