URL anti-spoofing.  A couple weeks late and completely untested, sorry.
The purpose of this is to avoid confusion caused by readers posting
what looks like a useful link but really goes somewhere misleading.

When a user saves a comment, this new code converts all "</A>" tags
to e.g. "</A abc.com>", where "abc.com" is the hostname given in the
corresponding <A> tag.  (Depends on balanceTags() having been run, and
also does some checking of the <A> tags on its own.  If a hostname is
not available, does its best to display something appropriate.)

Then when a user displays a comment, the illegal HTML code is converted
either to "</A> [abc.com]" or is simply removed to "</A>" depending on
the viewing user's wishes and whether the posting user, at that moment,
is allowing his/her email address to be shown.