For better or for worse, this is the real-email registration feature, plus a few minor additions.
This has been partially tested, but this check-in is already late and
it's damned near impossible to fully test this with the Messages
system still in flux. I'll have to hold off the rest of the testing
until integration time.

The system should at least work with these changes in. A lot of these
are new/changed templates and a new script, plus the tasks script which
replaces changes made to sbin/dailyStuff.

If there are any problems, please let me know.

- Cliff