Changes to slashcode-dearchives aren't even worth mentioning.
This is basically a commit of the first draft of the slash1toslash2.2
script which will be used to port Slashdot to fry.

THIS IS UNTESTED CODE! It's just here so that the rest of the team can
verify the current progress. Feel free to test it on your data, but
DO NOT use this in production at this time.