Merged changed code and users.pl  -- Hope I got it right, if not please flog me relentlessly.
Rather difficult to merge this because I had changes in my code that
interfered with PatG's, however the intent of this change is to make
SURE that the changes to the user form are properly reflected on
Banjo. Mostly it's all cosmetic changes to the code, but I hope I didn't
forget a line or two.