This is new code to cap comment_text caching more successfully, and to purge the cache more efficiently.  I'm making it live shortly.
The penalty is a slight overall speed decrease in raw comment
retrieval from cache.  About 5-10%.  This will add about 1 ms of
CPU to the average comment fetch.

I believe this will easily be offset by much better performance
when the cache approaches its maximum size.

The old code is still there.  If the var "comment_cache_newstyle"
is true, the new code will be used.  If it's 0, the old code will
be used, identical in almost every respect except that it'll be a
little smarter about caching.

Almost all the new code is in MySQL.pm:  _getCommentTextNew() and
_purgeCommentTextIfNecessary().

If the var "comment_cache_debug" is set to 1, some info on timing
and cache hit rates is dumped to STDERR.  If set to 2, the new
style code dumps a lot of debugging info to STDERR.  Also, there
will be overall cache hit and timing status at the top of /86.pl.