First phase of converting the four types of data uid, ipid, subnetid and classbid into a single data object: srcid. Ultimately this will also replace the 32-char MD5 strings stored in accesslog, comments, formkeys, moderatorlog, open_proxies, submissions, users_logtokens, and email_optout. But for now the only change is to the accesslist table.
The API of setAccessList, changeAccessList, checkReadOnly, getBanList,
and 9 other methods is still in place but will be deleted soon. It is
replaced by the API to the al2 tables: setAL2, createAL2Log, getAL2,
getAL2Log, checkAL2, getAL2List, getAL2Types, and getAL2TypeByID.

Furthermore, the user interface to getUserAdmin, in users.pl, now shows
a unified interface for both ACLs and access modifiers, as well as a
logging history of comments instead of a single "reason" field.