Nodnix autocomplete now handles admin commands as well as tags
Within createTag, now choose the admin op if the user is one,
this should do the right thing if the command turns out to just be
a simple tag.  Also updated the KeyListener to open editing for
admin command characters, and _not_ start up for the space key
any more (don't want to steal that from page navigation).