Use '-' instead of ':' in HTML ids built from sids
We don't want to use the original '/'s, because they are not legal characters in HTML ids
(see <http://www.w3.org/TR/html401/types.html>).  We don't want to use ':' or '.' because,
while they are legal, they are also meaningful in CSS rules, and complicate the transition
from prototype.js to jQuery.