Adding the core functionality of a tag bar
Most of the functionality is bundled together in tbar_fns.
The function tag_bar creates a new dom object for the bar
and extends it with the contents of tbar_fns.  Afterwards,
tag bar operations are called directly on the dom object
representing the tag bar.

This code is not yet included in all.js or any of the html
headers.