Remove unused expensive query from Firehose Usage console box
A query to get the total number of users who'd nodded and nixed was doing
a table scan.  Since its result wasn't being used in the template that
generates the Firehose Usage console box, I commented it out.  I also
moved all the other stats queries it does use to a reader instead of
the writer, and cleaned up some minor syntax.