Auto-sanitize submission data for authors/editors
The introtext is checked for a single case where it contains only one set of
double quotes and then if present they are replaced with single quotes.
Additionally, the subject line is sanitized so that every word is capitalized
save for a, an, and, the, & of. The first word of the submission is
capitalized regardless.