No wide characters in compress
Compress::Zlib now dies on wide characters.  Presumably it has a good reason.
We don't anticipate tons of wide characters, so we just convert them all to _.
If this is a problem later we can adjust filters accordingly, or something.
We could also perhaps use the utf8 -> ascii filters in the same module.