Tag-widget: add input field, animate open/close, add spinner, fix css
Added a text field (in a form), and for better css control wrapped the tag bars
themselves in a div of their own.  Added css to support both "stack" and "train"
layouts of the bars.  We are currently using the "train" (one-line) layout.

_submit_fetch core starts the busy spinner on entry, stops it after all work from
the completion function is done.

For paranoia, handle a possible case in the firehose vote handler, where Jamie
tells me both nod and nix.

Added some debugging css to show Rob the top and system tag-bars.