Adding new JS function firehose_id_of(expr) to common.js
expr can be a number, a dom element that is or is within a firehose entry, or a
string that is a number or the id of a dom element that is or is within a
firehose entry (not just any id, though --- it must be an id that contains the
firehose id number as a suffix).  Returns the integer id, if we find one, or
else undefined.

Useful in handlers on dom elements within a firehose entry, e.g.,

	$('.twisty').click(function(){
		// 'this' now refers to the dom element accepting the event
		tagsHideBody(firehose_id_of(this))
	})