Teach a tag-widget how to do one unified fetch instead of three little ones
The unified answer will come back looking like this:

	<user>tag1 tag2 tag3\n<top>tag1 tag2 tag3\n<system>tag1 tag2 tag3

So we give the widget its own set_tags() function which parses that unifed
response and pushes the pieces into the appropriate child tag-bar.

widget.fetch_tags() and bar.fetch_tags() look more alike now, so clean-up the
latter a little to match.

Now that the widget itself is fetching, it also stores the firehose id of the
entry to which it is attached.