Teach setTagsForGlobj a new way to deactivate; new ajax_op to use it
setTagsForGlobj takes as input a mixed list of tags and admin commands.  It
sends the admin commands off to be processed elsewhere.  Before this commit,
whatever was left in that list specified exactly what you wanted your user tags
to be (for some particular item). Any tag in that list you _don't_ have, gets
added.  If it's different in the list from your existing tag, e.g., tag vs !tag,
your tag gets updated.  And (here's the biggy) any tag you have currently that
_does_not_ appear in that list gets deactivated (deactivation by omission).

This commit teaches setTagsForGlobj that when called with the option
'deactivate_by_operator', it must not touch any tags not explicitly mentioned in
the input.  A new user-level command is introduced for deactivation: -tag
(deactivation by commission).  With this change, setTagForGlobj becomes a single
point of entry for operating on a user's tags on a given item: callers get all
its original functionality, but without first querying for current tags.

ajaxSetGetCombined then wraps that functionality and exposes it to ajax, calling
setTagsForGlobj if needed, and then always returning a combined list of
{user,top,system} tags.

Tags.pm:
	+ teach deactivateTag that sometimes the tagname _is_ required
	+ teach setTagsForGlobj an alternative protocol for deactivating tags
	+ add ajaxSetGetCombinedTags

upgrades, mysql_dump.sql:
	+ tell the database about our new ajax_op, tags_setget_combined