Fields find the widget, not the reverse; rename one of the normalize functions
There's no reason to make the widget know anything specific about the input
field(s); in fact, it's more efficient if the input field knows directly how to
call the widget's submit_tags function --- because then the widget doesn't have
to search for the input field.  Made this easy to say in the HTML for the field
by factoring it into a global function: textfield_submit_tags().

Renamed normalize_tag_command to normalize_tag_menu_command for clarity.