Hook jQuery's autocompleter into the tag-widget's busy feedback
We want autocomplete to show the same `busy' feedback while fetching completions
that we already show while setting and fetching tags.  Unfortunately, jQuery's
autocompleter has only one means of showing this feedback: when it starts an
ajax request, it sets a class ('ac_loading' by default) on the input field; on
completion, it removes the class.  Because we're not allowed (HTML law) to put
our feedback element logically inside the input element, we make it be the
immediate next sibling of the input field and use the CSS adjacent-sibling
selector, '+', to manage its visibility.

Happily, jQuery's show() and hide() still work so no changes are needed to the
existing busy code.