As with tag-widgets, allow tag-bars to be stubs; saving code in favor of templates
The stub-expanding code decides you are a stubbed tag-bar if
	(a) you say so, class="tbar stub"
	(b) you have a 'get' attribute and _no_ class attribute

Sample tag-bar stubs:

	<div class="tbar stub" get="user" label="My Tags">tag1 tag2 tag3</div>
	<div get="top"></div>

The stub class will be deleted, and 'tbar' will be added if not present.  The
value of your label attribute will become the content in an <h1 class="legend">
element; the attribute itself will be deleted.  A menu attribute can override
the default tag menus; the attribute itself will be deleted.

	+ tagbar_data.list_el, the saved reference to the <ul> is now a jQuery
	  object, so renamed $list_el
	+ twidget_fns.init: no longer adds specific bars, now just expands stubs
	+ new init_tag_bar(): expands and initializes a stubbed tag-bar
	+ rewrote create_tag_bar() to now be a wrapper around init_tag_bar
	+ added stub tag-bars to the widget template; we are no longer
	  hand-coding them
	+ create_firehose_vote_handler: assigns a class to prevent its listener
	  from being treated like a stub