Teach the tag widget to style tags with global observations
E.g., if a tag appears both in the user tag bar and the top tags bar, then every
occurance of the tag anywhere in that widget should be styled to reflect that.
With that knowledge, the css rules could (for instance) hide in the user bar any
tags that also appear in the top tags bar.

Also made the style code smarter about `well known tags', though this machinery
needs to move into a template.

Big changes in the css to (a) make my rules win over more general base rules,
(b) make the tag menus at least visible if not usable, (c) hide tags, as above,
(d) renaming the prefix styles, (e) adding new styles for some of the well known
tags.