Sort hof.pl "Active Authors" correctly
An unnecessary 'GROUP BY uid' was causing the ORDER BY clause to
be ignored -- presumably having something to do with the query's
non-aggregate columns being technically illegal (the only reason
MySQL didn't throw an error was that ONLY_FULL_GROUP_BY mode was
not on).  Removing the GROUP BY allows MySQL to sort by the
desired column.
<http://dev.mysql.com/doc/refman/5.0/en/group-by-hidden-fields.html>
<http://dev.mysql.com/doc/refman/5.0/en/server-sql-mode.html>