The nod/nix reasons widget _doesn't_ have to be a special case
Tag displays now notice when they become empty/non-empty, and mark themselves
with the CSS class 'no-tags', accordingly.  CSS rules can now hide them.  When
the user nods/nixs, the context is set and the reasons appear.  When the user
clicks a reason, the context is set again, but this time to something that
empties the nod/nix reasons display... and because it's empty, it immediately
gets marked as such, and the CSS hides it.

Tried to pick some more sensible names for some things.