A smarter tag command `normalizer'
The old normalizer was used to perform three fixes to a string of tag commands:
- s/!nod/nix/g; s/!nix/nod/g (as nod and nix are already opposites of each
  other, and aren't set up to be opposites in the '!' way)
- any occurance of 'nod' becomes 'nod -nix' (similarly when starting with 'nix')
  to handle a funny situation where---believe it or not---an item can be both
  nodded and nixed at once
- for feedback only, to remove nod/nix (of every spelling) from the string to
  prevent it from showing up in the user tag display

This commit adds additional optimization tasks:
- remove duplicates or opposites, yields minmal sequence for the same result
- optionally filter against your current tags (so you don't try to add again, or
  to deactivate a tag you don't actually have)
- this means an entire command sequence could be "optimized away", so now check
  for that and don't bother sending an ajax request

normalize_tag_command() is fairly heavily commented, so it looks bulkier than it
is.  The meat is in the $.map near the end.  Everything above that is really
just about massaging the input parameters.

Also added: cached_user_tags(), which finds a descendent user tag display and
extracts those tags (along with a 'nod' or a 'nix' if you've voted).  This is
how normalize_tag_commands() gets a list to filter against if you passed in a
jQuery selector for its second param.