Teach nearest_parent to provide nearest parents for _each_ wrapped element
nearest_parent (lacking a previous description in commit messages) finds the
nearest enclosing dom element (including the starting element) that matches the
provided selector.  For example, to find the containing firehose entry from a
click handler:

  $(event.target).nearest_parent('[id^=firehose-]')

That works even if the event target happened to be the firehose entry itself.

To find the clicked tag:

  $(event.target).nearest_parent(':has(span.tag)')

I use this function everywhere to get to the DOM element that has the tag ajax
code (usually the firehose entry itself), e.g.,

  $(this).nearest_parent('[tag-server]')

This improvement means that you could find a minimal set of parents across a
collection of elements, e.g., get all the firehose entries associated with the
elements you're currently holding, even when elements share the same entry as an
ancestor.

- ':first' may provide better performance than .eq(0)
- this.map lets me handle wrapped elements individually
- $.unique makes the answer minimal
- this.pushStack makes .end() do the right thing after using nearest_parent