Use ':first' instead of .eq(0) when reasonable
With $(expr).eq(0), you've already done the work (e.g., finding 200 different
DOM elements).  With $('expr:first'), jQuery at least has the _opportunity_ to
stop the search immediately.  I don't know that it _does_, but if not now it may
in the future.  And perhaps $('.article:first') is more obvious to the reader
than $('.article').eq(0).