Speed up listTagnamesRecent
listTagnamesRecent is used by /tags/recent and the console's Recent
Tagnames box.  When tag creation (esp. with many unique tagnames)
was briefly increased to on the order of 5/sec over 12 hours, its
main SELECT was unworkably slow.  This should get it back to
responding in 1 second or less even in such cases.