FireHose memcached bugfix
If firehose_mcd_disp was set to false, in some cases, firehose data
was not being properly expired even for non-display uses.  This was
causing problems for the tagbox that set a hose item to is_spam=yes
and then immediately recalculated its score:  the old is_spam=no
value was being retrieved from memcached so its score did not change.

Also, moved a hardcoded expiration time into a var.