Slash.ArticleInfo: store (in a universal format) article-related data needed by client code
This builds on the work pudge and I did for sfnet.  'Article', here, is a general name for
an entry in the main list of the page, e.g., a firehose entry, or the story in editStory,
etc.  Uses only valid HTML.  You hang your sd-info-block directly from the element that
represents the article, e.g.,

<div id="firehose-1128650" class="briefarticle adminmode" ... >
	<span class="sd-info-block">
		<span class="sd-key-firehose-id">1128650</span>
		<span class="type">firehose</span>
		<span class="new-article">true</span>
	</span>

	...

</div>

$('#firehose-1128650').article_info('type') => 'firehose'
$('#firehose-1128650').article_info('init-time', new Date());

If you can't hang it right from the article element, provide a 'scope' definition:

	<span class="sd-info-block">
		<span class="sd-key-firehose-id">1128650</span>
		<span class="scope">#firehose-1128650</span>
		<span class="type">firehose</span>
	</span>

Given any element that is or is within an article that has an sd-info-block

	$(elem).article_info__find_articles()

finds that outer article-element.