Floating Slashbox Ad for the firehose
This commit adds the package Slash.Firehose.floating_slashbox_ad, and two extra methods
to Slash.Firehose itself: choose_article_for_next_ad(), and articles_on_screen().

Use it like this:

	// Let's make a short-cut.
	var floating_ad = Slash.Firehose.floating_slashbox_ad;

	...

	if ( ! floating_ad.is_visible() ) {
		var $article = Slash.Firehose.choose_article_for_next_ad();

		... get an iframe (or other ad content) from the ad server ...

		floating_ad(ad_content, $article);
	}

The ad will be top-aligned with the article, but pinned to the bounds of the window
and the empty space in the slashboxes column.  When the attached article is off-screen,
the ad will be hidden.  If the attached article is scrolled back on-screen, the ad will
be re-shown.

If no article is provided (or if you call floating_ad.detach()), then the ad is not
constrained by an article.  Setting an ad, as above, automatically clears the
current ad; but you can do this manually as well: floating_ad.remove().  The ad position
is 300px x 300px, unless you call floating_ad.set_size(new_height, new_width).  The
new height must be an integer number of pixels, but the new width can be symbolic, e.g.,
'100%'.  You can recover the article to which the current ad is attached with
floating_ad.current_article().

Slash.Firehose.articles_on_screen() returns a jQuery selection of all the articles that
are currently visible and intersect the window bounds (even by 1px).  choose_article_for_next_ad()
calls articles_on_screen() and then randomly selects an element from the returned list.

Probably want to add code to the slashbox machinery to ensure the sentinel introduced here
remains last in the list.