Fix __SLASHLINK__ admin display bug
Links to (the same) Slash site in the bodytext of a story were
showing up as their unparsed __SLASHLINK__ text, when viewed in
article.pl, admin.pl, etc., by an admin.  This was because admins
triggered processing of the neverdisplay template, which overwrote
the (correctly parsed) data in $story between the display of the
introtext and the bodytext.  Fixed by having the neverdisplay
template use a global variable name that's less likely to step on
other templates' variables.