Always link to real_rootdir for /~ URLs
Previously, we linked to URLs whose path began with /~ either with
the raw path (which implied the current hostname) or with gSkin.rootdir
(which implied the current hostname normalized) or with constants.rootdir
(which I believe may vary based on current skin, though I'm not
sure of the exact circumstances).  Now we use constants.real_rootdir
consistently, which always uses the mainpage skin's hostname.